Tweak: Add `useStyleIndicator` hook
