+++ src/utils/get-content-attribute/index.js
@@ -1,0 +1,13 @@
+export function getContentAttribute( attributes, blockName ) {
+	let contentValue = '';
+
+	if ( 'generateblocks/button' === blockName ) {
+		contentValue = attributes.text;
+	}
+
+	if ( 'generateblocks/headline' === blockName ) {
+		contentValue = attributes.content;
+	}
+
+	return contentValue;
+}
